Extend RPi fb layer driver to provide triple-buffering
This commit extends the fb layer driver used by the Raspberry Pi in
drivers/video/fbdev/bcm2708_fb.c to provide three screen pages,
allowing double or triple-buffering and facilitating smooth and
tear-free animation for console applications that do not use
EGL/GLES2. The size of the framebuffer memory region visible to
console applications is now three times as large as it used to be.

The DMA-accelerated CopyArea function that was already provided by the
fb layer is also extended to allow access to all three screen pages.
In order to do so, the function fb_copyarea_user in the core fb layer
is modified to allow access to the framebuffer beyond the first screen.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>